+++ src/cmd/prepend.go
@@ -33,26 +33,12 @@ This can be disabled by toggling the "new-branch-push-flag" configuration:
 
 	git town new-branch-push-flag false`,
 	Run: func(cmd *cobra.Command, args []string) {
-		steps.Run(steps.RunOptions{
-			CanSkip:              func() bool { return false },
-			Command:              "prepend",
-			IsAbort:              abortFlag,
-			IsContinue:           continueFlag,
-			IsSkip:               false,
-			IsUndo:               undoFlag,
-			SkipMessageGenerator: func() string { return "" },
-			StepListGenerator: func() steps.StepList {
-				config := getPrependConfig(args)
-				return getPrependStepList(config)
-			},
-		})
+		config := getPrependConfig(args)
+		stepList := getPrependStepList(config)
+		runState := steps.NewRunState("prepend", stepList)
+		steps.Run(runState)
 	},
-	Args: func(cmd *cobra.Command, args []string) error {
-		if abortFlag || continueFlag || undoFlag {
-			return cobra.NoArgs(cmd, args)
-		}
-		return cobra.ExactArgs(1)(cmd, args)
-	},
+	Args: cobra.ExactArgs(1),
 	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return util.FirstError(
 			git.ValidateIsRepository,
@@ -90,8 +76,5 @@ func getPrependStepList(config prependConfig) (result steps.StepList) {
 }
 
 func init() {
-	prependCommand.Flags().BoolVar(&abortFlag, "abort", false, abortFlagDescription)
-	prependCommand.Flags().BoolVar(&continueFlag, "continue", false, continueFlagDescription)
-	prependCommand.Flags().BoolVar(&undoFlag, "undo", false, undoFlagDescription)
 	RootCmd.AddCommand(prependCommand)
 }