+++ src/steps/run_state.go
@@ -1,17 +1,37 @@
 package steps
 
-import "github.com/Originate/git-town/src/git"
+import (
+	"time"
+
+	"github.com/Originate/git-town/src/git"
+)
+
+// UnfinishedRunStateDetails has details about an unfinished run state
+type UnfinishedRunStateDetails struct {
+	CanSkip   bool
+	EndBranch string
+	EndTime   time.Time
+}
 
 // RunState represents the current state of a Git Town command,
 // including which operations are left to do,
 // and how to undo what has ben done so far.
 type RunState struct {
-	AbortStep    Step
-	Command      string
-	IsAbort      bool
-	isUndo       bool
-	RunStepList  StepList
-	UndoStepList StepList
+	AbortStepList     StepList
+	Command           string
+	IsAbort           bool
+	isUndo            bool
+	UnfinishedDetails *UnfinishedRunStateDetails
+	RunStepList       StepList
+	UndoStepList      StepList
+}
+
+// NewRunState returns a new run state
+func NewRunState(command string, stepList StepList) *RunState {
+	return &RunState{
+		Command:     command,
+		RunStepList: stepList,
+	}
 }
 
 // AddPushBranchStepAfterCurrentBranchSteps inserts a PushBranchStep
@@ -36,7 +56,7 @@ func (runState *RunState) AddPushBranchStepAfterCurrentBranchSteps() {
 func (runState *RunState) CreateAbortRunState() (result RunState) {
 	result.Command = runState.Command
 	result.IsAbort = true
-	result.RunStepList.Append(runState.AbortStep)
+	result.RunStepList.AppendList(runState.AbortStepList)
 	result.RunStepList.AppendList(runState.UndoStepList)
 	return
 }
@@ -45,7 +65,7 @@ func (runState *RunState) CreateAbortRunState() (result RunState) {
 // that skips operations for the current branch.
 func (runState *RunState) CreateSkipRunState() (result RunState) {
 	result.Command = runState.Command
-	result.RunStepList.Append(runState.AbortStep)
+	result.RunStepList.AppendList(runState.AbortStepList)
 	for _, step := range runState.UndoStepList.List {
 		if isCheckoutBranchStep(step) {
 			break
@@ -72,6 +92,25 @@ func (runState *RunState) CreateUndoRunState() (result RunState) {
 	result.isUndo = true
 	result.RunStepList.AppendList(runState.UndoStepList)
 	return
+}
+
+// IsUnfinished returns whether or not the run state is unfinished
+func (runState *RunState) IsUnfinished() bool {
+	return runState.UnfinishedDetails != nil
+}
+
+// MarkAsFinished updates the run state to be marked as finished
+func (runState *RunState) MarkAsFinished() {
+	runState.UnfinishedDetails = nil
+}
+
+// MarkAsUnfinished updates the run state to be marked as unfinished and populates informational fields
+func (runState *RunState) MarkAsUnfinished() {
+	runState.UnfinishedDetails = &UnfinishedRunStateDetails{
+		CanSkip:   false,
+		EndBranch: git.GetCurrentBranchName(),
+		EndTime:   time.Now(),
+	}
 }
 
 // SkipCurrentBranchSteps removes the steps for the current branch